% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stir_ER.R
\name{stir_ER}
\alias{stir_ER}
\title{Stir an Erdos-Renyi Random Network}
\usage{
stir_ER(matrix, x, p, directed = TRUE, retcon = FALSE)
}
\arguments{
\item{matrix}{Existing network to experience growth.}

\item{x}{The ID of the node to be rewired (stirred).}

\item{p}{Probability possible edges exist. Needs to be between zero and one.}

\item{directed}{Binary variable determining if the network is directed, resulting in off-diagonal asymmetry in the adjacency matrix.}

\item{retcon}{Binary variable determining if already existing nodes can attach to new nodes. Defaults to FALSE.}
}
\value{
An adjacency matrix.
}
\description{
Stirs an already existing network by rewiring a node according to the Erdos-Renyi random mechanism.
}
\details{
Different from Duplication & Mutation models in that edges can only be lost.
}
\examples{
# Import netcom
library(netcom)

size <- 10
existing_network <- matrix(sample(c(0,1), size = size^2, replace = TRUE), nrow = size, ncol = size)
new_network_prep <- matrix(0, nrow = size + 1, ncol = size + 1)
new_network_prep[1:size, 1:size] = existing_network
new_network <- stir_ER(matrix = new_network_prep, x = size + 1, p = 0.5)

}
\references{
Erdos, P. and Renyi, A., On random graphs, Publicationes Mathematicae 6, 290–297 (1959).
}
