% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_test.R
\name{unit_test}
\alias{unit_test}
\title{Test of the align.R function}
\usage{
unit_test()
}
\value{
Either "Unit test results: PASS" or "Unit test results: FAIL"
}
\description{
Compared the output of align.R to a hand curated output.
}
\examples{
unit_test()
netcom::unit_test()

}
