% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gini.R
\name{gini}
\alias{gini}
\title{Gini coefficient}
\usage{
gini(input)
}
\arguments{
\item{input}{A row-stochastic matrix, meaning each row is a multinomial distribution.}
}
\value{
A vector of the Gini coefficient of each row.
}
\description{
Takes a matrix and returns the Gini coefficient of each row.
}
\examples{
gini(matrix(runif(25, 0, 1), nrow = 5, ncol = 5))
netcom::gini(matrix(runif(25, 0, 1), nrow = 5, ncol = 5))

}
