% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ics.R
\name{ics}
\alias{ics}
\title{Induced Conserved Structure (ICS)}
\usage{
ics(matrix_1_input, matrix_2_input, alignment, input = "matrix",
  flip = FALSE)
}
\arguments{
\item{matrix_1_input}{The first network being aligned, either as a matrix or linked list. If the two
networks are of different sizes, it will be easier to interpret the output if this is the smaller one.}

\item{matrix_2_input}{The second network. Should be the same type (matrix or linked list) as matrix_1_input.}

\item{alignment}{A matrix, such as is output by the function NetCom, where the first two columns contain
corresponding node IDs for the two networks that were aligned.}

\item{input}{Defaults to "matrix". Can be set to "list" if the two networks are stored as linked lists.}

\item{flip}{Defaults to FALSE. Set to TRUE if the first network is larger than the second. This is necessary 
because ICS is not a symmetric measure of alignment quality.}
}
\value{
A number ranging between 0 and 1. If the Induced Conserved Structure is 1, the two networks are
isomorphic (identical) under the given alignment.
}
\description{
Calculates the Induced Conserved Structure (Patro and Kingsford, 2012) of an alignment between two networks.
}
\examples{
# Note that ICS is only defined on unweighted networks.
net_one <- round(matrix(runif(25,0,1), nrow=5, ncol=5))
net_two <- round(matrix(runif(25,0,1), nrow=5, ncol=5))
ics(net_two, net_two, align(net_one, net_two)$alignment)

}
