% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_DM.R
\name{make_DM}
\alias{make_DM}
\title{Make a Duplication and Mutation Network}
\usage{
make_DM(size, net_kind, divergence, mutation, directed = FALSE)
}
\arguments{
\item{size}{Number of nodes in the network.}

\item{net_kind}{If the network is an adjacency matrix ("matrix") or an edge list ("list").}

\item{divergence}{Probability that the new node loses edges associated with the node it duplicates. Needs to be between zero and one.}

\item{mutation}{Probability that the new node gains edges not associated with the node it duplicates. Needs to be between zero and one.}

\item{directed}{Binary variable determining if the network is directed, resulting in off-diagonal asymmetry in the adjacency matrix. Defaults to TRUE.}
}
\value{
An adjacency matrix.
}
\description{
Make an already existing network according to the Duplication and Mutation mechanism.
}
\details{
Different from Duplication & Mutation models in that edges can only be lost.
}
\examples{
size <- 10
existing_network <- matrix(sample(c(0,1), size = size^2, replace = TRUE), nrow = size, ncol = size)
new_network_prep <- matrix(0, nrow = size + 1, ncol = size + 1)
new_network_prep[1:size, 1:size] = existing_network
new_network <- make_DM(matrix = new_network_prep, x = size + 1, divergence = 0.5)

}
\references{
Ispolatov, I., Krapivsky, P. L., & Yuryev, A. (2005). Duplication-divergence model of protein interaction network. Physical review E, 71(6), 061911.
}
