% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_NM.R
\name{make_NM}
\alias{make_NM}
\title{Make a Niche Model network}
\usage{
make_NM(size, niches, connectance = 0.5, directed = TRUE)
}
\arguments{
\item{size}{The number of nodes in the network. Must be a positive integer.}

\item{niches}{A vector of numbers specifying the niche of each member of the system (node). Each niche value must be element of [0,1].}

\item{connectance}{Defaults to 0.5. The ratio of actual interactions to possible interactions. Effects the beta distributed width of niche values each member of the system (node) interacts with.}

\item{directed}{Defaults to TRUE. If FALSE all interactions will be made symmetric. Note that the process of creating interactions is unaffected by this choice.}
}
\value{
An interaction matrix format of a Niche Model network.
}
\description{
Creates a single network according to the Niche Model. Can be directed or undirected, but is always unweighted.
}
\examples{
# Network size (number of nodes)
size <- 10

# Create niche values for each member of the system (node)
niches <- runif(n = size)

# Make network according to the Niche Model
make_NM(size = size, niches = niches)

}
\references{
Williams, R. J., & Martinez, N. D. (2000). Simple rules yield complex food webs. Nature, 404(6774), 180-183.
}
