% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_SW.R
\name{make_SW}
\alias{make_SW}
\title{Makes a Small-World Network}
\usage{
make_SW(size, net_kind, rewire, neighborhood, directed = FALSE)
}
\arguments{
\item{size}{The number of nodes in the network. Must be a positive integer.}

\item{net_kind}{The format of the network. Currently must be either `matrix` or `list`.x}

\item{rewire}{Small-World parameter specifying the probability each edge is randomly rewired, allowing for the possiblity of bridges between connected communities.}

\item{neighborhood}{The range of nodes that form connected communities. Note: This implementation results in overlap of communities.}

\item{directed}{Binary variable determining if the network is directed, resulting in off-diagonal asymmetry in the adjacency matrix. Defaults to TRUE.}
}
\value{
An adjacency matrix.
}
\description{
Make an already existing network according to the Small-World mechanism.
}
\details{
Rewires a node in a network according to the Small-World mechanism.
}
\examples{
size <- 10
existing_network <- matrix(sample(c(0,1), size = size^2, replace = TRUE), nrow = size, ncol = size)
new_network_prep <- matrix(0, nrow = size + 1, ncol = size + 1)
new_network_prep[1:size, 1:size] = existing_network
new_network <- make_SW(matrix = new_network_prep, x = size + 1, rewire = 0.213)

}
\references{
Watts, D. J., & Strogatz, S. H. (1998). Collective dynamics of ‘small-world’networks. nature, 393(6684), 440-442.
}
